create the EmailForm folder and break up the individual fields into sub-components so that they can be re-used in different forms
