+++ src/components/Download/EmailForm/ReceiveUpdatesCheckBox.js
@@ -1,0 +1,14 @@
+import { CheckBox } from 'grommet'
+
+export const ReceiveUpdatesCheckBox = ({ value, handleChange }) => {
+  return (
+    <CheckBox
+      checked={value}
+      label="I would like to receive occasional updates from the refine.bio team"
+      name="receiveUpdates"
+      onChange={handleChange}
+    />
+  )
+}
+
+export default ReceiveUpdatesCheckBox
